Refactor sensuctl create and process functions for generic use (#3616)
