+++ cli/commands/asset/add.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cobra"
 
 	goversion "github.com/hashicorp/go-version"
-	"github.com/sensu/sensu-go/cli/commands/create"
+	"github.com/sensu/sensu-go/cli/resource"
 )
 
 var rename string
@@ -73,11 +73,11 @@ func addCommandExecute(cli *cli.SensuCli) func(cmd *cobra.Command, args []string
 			return err
 		}
 
-		resources, err := create.ParseResources(bytes.NewReader([]byte(asset)))
+		resources, err := resource.Parse(bytes.NewReader([]byte(asset)))
 		if err != nil {
 			return err
 		}
-		if err := create.ValidateResources(resources, cli.Config.Namespace()); err != nil {
+		if err := resource.Validate(resources, cli.Config.Namespace()); err != nil {
 			return err
 		}
 		for i := range resources {
@@ -89,7 +89,8 @@ func addCommandExecute(cli *cli.SensuCli) func(cmd *cobra.Command, args []string
 			}
 			resources[i].Value.SetObjectMeta(meta)
 		}
-		if err := create.PutResources(cli.Client, resources); err != nil {
+		processor := resource.NewPutter()
+		if err := processor.Process(cli.Client, resources); err != nil {
 			return err
 		}
